DX9/DX11: Fixing some maybe possible crashes if a game was started, the config dialog opened and the game closed again. Due to other issues this still happens quite often though... Various warning fixes.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@6684 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Plugins/Plugin_VideoDX11/Src/D3DBlob.cpp
@@ -27,7 +27,7 @@ D3DBlob::D3DBlob(ID3D10Blob* d3dblob) : ref(1)
 {
 	blob = d3dblob;
 	data = (u8*)blob->GetBufferPointer();
-	size = blob->GetBufferSize();
+	size = (unsigned int)blob->GetBufferSize();
 	d3dblob->AddRef();
 }
